+++ android/lib/common/src/main/kotlin/net/mullvad/mullvadvpn/lib/common/util/ErrorNotificationMessage.kt
@@ -1,7 +1,0 @@
-package net.mullvad.mullvadvpn.lib.common.util
-
-data class ErrorNotificationMessage(
-    val titleResourceId: Int,
-    val messageResourceId: Int,
-    val optionalMessageArgument: String? = null,
-)