+++ android/lib/model/src/main/kotlin/net/mullvad/mullvadvpn/lib/model/ConnectError.kt
@@ -3,5 +3,5 @@ package net.mullvad.mullvadvpn.lib.model
 sealed interface ConnectError {
     data class Unknown(val throwable: Throwable) : ConnectError
 
-    data object NoVpnPermission : ConnectError
+    data class NotPrepared(val error: PrepareError) : ConnectError
 }